add existence check for /etc/apt/source.list

+++ misc/change_apt_source.sh
@@ -2,6 +2,8 @@
 set -e
 set -x
 
-mv /etc/apt/sources.list /etc/apt/sources.list.bak
+if [ -f /etc/apt/sources.list ]; then
+    mv /etc/apt/sources.list /etc/apt/sources.list.bak
+fi
 cp ./sources/sources.list /etc/apt/sources.list
 apt-get update
